fix testapplication bug

+++ app/src/androidTest/java/com/faz/stormtroopers/test/factory/TripFactory.kt
@@ -31,7 +31,7 @@ class TripFactory {
             return TripModel.Pilot(
                 DataFactory.randomUuid(),
                 DataFactory.randomUuid(),
-                DataFactory.randomInt()
+                DataFactory.randomFloat()
             )
         }
